requirements.txt
# 基础依赖
ccxt>=4.0.0          # 统一交易所 API
pyyaml>=6.0          # 配置文件解析
python-dotenv>=1.0   # 环境变量管理

# 数据处理
pandas>=2.0.0        # 数据处理
numpy>=1.24.0        # 数值计算

# 异步支持
aiohttp>=3.8.0       # 异步 HTTP
asyncio-throttle>=1.0  # API 限流

# 日志与通知
loguru>=0.7.0        # 日志系统
python-telegram-bot>=20.0  # Telegram 通知

# 开发工具
pytest>=7.0.0        # 测试框架
pytest-asyncio>=0.21  # 异步测试

# 可视化（可选）
plotly>=5.0.0        # 交互式图表

